+++ modules/app/src/main/scala/com/fortysevendeg/ninecardslauncher/process/cloud/CloudStorageProcess.scala
@@ -1,7 +1,7 @@
-package com.fortysevendeg.ninecardslauncher.process.drive
+package com.fortysevendeg.ninecardslauncher.process.cloud
 
 import com.fortysevendeg.ninecardslauncher.commons.services.Service.ServiceDef2
-import com.fortysevendeg.ninecardslauncher.process.drive.models.{CloudStorageDevice, CloudStorageResource}
+import com.fortysevendeg.ninecardslauncher.process.cloud.models.{CloudStorageDevice, CloudStorageResource}
 
 trait CloudStorageProcess {
