+++ modules/app/src/main/scala/com/fortysevendeg/ninecardslauncher/process/cloud/Conversions.scala
@@ -1,6 +1,6 @@
-package com.fortysevendeg.ninecardslauncher.process.drive
+package com.fortysevendeg.ninecardslauncher.process.cloud
 
-import com.fortysevendeg.ninecardslauncher.process.drive.models.CloudStorageResource
+import com.fortysevendeg.ninecardslauncher.process.cloud.models.CloudStorageResource
 import com.fortysevendeg.ninecardslauncher.services.drive.models.DriveServiceFile
 
 trait Conversions {
